requirements.txt
asgiref==3.6.0
attrs==22.2.0
coverage==7.2.2
Django==4.2
djangorestframework==3.14.0
exceptiongroup==1.1.1
factory-boy==3.2.1
Faker==18.3.2
inflection==0.5.1
iniconfig==2.0.0
model-bakery==1.10.3
packaging==23.0
pluggy==1.0.0
psycopg2-binary==2.9.6
pytest==7.2.2
pytest-cov==4.0.0
pytest-django==4.5.2
pytest-factoryboy==2.5.1
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.5.0
